main1-main/backend/requirements.txt
Flask==2.3.3
fpdf2==2.7.5
qrcode[pil]==7.4.2
Pillow==10.0.0
